requirements.txt
celery==5.5.3
Django==5.2.1
weasyprint==65.1
djangorestframework==3.16.0
chardet==5.2.0 
reportlab==4.4.1
cron-descriptor==1.4.5
django-celery-beat==2.8.1 
django-timezone-field==7.1 
python-crontab==3.2.0
django-cors-headers==4.7.0